Handle malformed shebangs
One of the tests in valgrind specifically provides a shebang that is
malformed (contains just a `#!`), which is technically valid bourne
shell syntax.

Fix the code to check that the result from `shlex.split(..)` isn't an
empty `tuple` to avoid an `IndexError` traceback when malformed input
is provided to `identify`.

Add a testcase for the issue.