imp: add nix-shell interpreter parser
nix-shell uses the second line to pass relevant arguments
to the interpreter. Linux (unlike BSD)
-- see: https://unix.stackexchange.com/a/63981 --
only passes a single argument to the shebang command
(in this case env or nix-shell).

On the second line, if specified, the interpreter is passed
to nix-shell via the -i flag, so we parse this interpreter
if the second line is specified. If not (nix-shell called w/o
arguments) the implicit interpreter is bash.
-- see also: #127 --