fix: make PKGBUILD files a superset of .bash
Arch Linux/ALPM PKGBUILDs are not currently being identified as `shell`
files, but they’re really just specialised bash scripts, and thus also
`shell` files.

This makes `PKGBUILD` files a superset of the `bash` extension, which
both makes it inherit the `shell` type, and also ensures that it will
automatically adopt any future changes/additions that might be made to
the `bash` extension.